pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-image-filter-20250329033722"
version = "0.1.0"
authors = [
  {name = "Your Name", email = "your.email@example.com"},
]
description = "Filter unrealistic images based on RGB histogram analysis"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "numpy>=1.19.0",
    "opencv-python>=4.5.0",
    "matplotlib>=3.3.0",
    "pillow>=8.0.0",
    "scipy>=1.6.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/simple-image-filter"
"Bug Tracker" = "https://github.com/yourusername/simple-image-filter/issues"

[project.scripts]
imgfilter = "simple_image_filter.cli:main"
